Allow for deploying into a specific namespace
This adds the target namepsace for the cluster-api components to the
operator config rather than deploying everything to the default
namespace.  Deploying to a different namespace requires different TLS
assets because the namespace is part of the service's DNS name in the
cluster.  Because of that, this also adds the aggregated API service
CA as part of the config, and removes the secret containing the TLS
assets from the manifests directory.  The OpenShift installer is
already generating these, so we should now be able to use the files
generated by the installer on real clusters, while still using the
pre-generated assets in the repo for development.