Add updateNodeRef for node link controller
Originally the logic for setting the nodeRef in a machine was split between the cluster-api node controller and opt-int external tooling that would set an annotation for the former controller to watch. There's no reason for this to keep leaving in two different places, furthermore the cluster-api node controller is deprecated. This unify this logic in one single nodelink controller