RBAC: Add additional metal3.io resource access
When utilising the baremetal infrastructure platform the baremetal
operator component needs to be able to update additional resources
for baremetal hosts, namely the status and the finalizer. This patch
ensures that it can do so, and will now allow hosts to be registered
with OpenShift as `BareMetalHosts` and as a result inside of Ironic.