Add baremetalhosts.metal3.io to related objects.
This patch is similar to PR #391, which added machines and
machinesets to the related objects list for the cluster operator.
This will ensure that BareMetalHosts are collected by must-gather, as
well.

This is only applicable when the "baremetal" infrastructure platform
is in use, but is harmless otherwise as the CRD is always defined.