baremetal: Add liveness probe to Ironic container
Without a liveness probe on the Ironic container, kubernetes never
seems to notice that the service is unavailable when the node goes
down. This causes recovery from an outage to take much longer than
it should and impacts the user experience. This change adds a basic
liveness probe which in my testing allows the outage to be recognized
in a reasonable amount of time. It limits the amount of time that the
Ironic API is unavailable.