add more FAQ information about annotations, labels, and taints
This change adds a new question and augments the current question on
labels. This is being proposed to fill out the user facing information
about how to properly apply annotations, labels, and taints to Machine
and Node objects. Some minor reformatting and coalescing has been done
to help maintain a consistent style.