pkg/operator/status: Change ReasonEmpty to ReasonAsExpected
We've used ReasonEmpty since 4000ff88f0 (Set all expected
ClusterOperator status conditions, 2019-01-10, #174).  But if we feel
like we have a message we want to set to help humans understand the
condition, we should be setting a reason string for machines
too.  AsExpected follows the library-go precedent [1].

Also use AsExpected for some message-less but expected conditions,
like Progressing=False and Degraded=False.  Having a reason makes it
easier for folks skimming logs to see that the condition is in its
happy state, and doesn't cost us any effort to set.

Also shuffle defaultStatusConditions a bit:

* Progressing=False didn't make sense to me as an initial condition.
  It is now:

    Progressing=True Initializing: Operator is initializing

  to show that the operator is working on getting itself ready.

* Available=False is a surprising condition and needs an explaination.
  Now we explain it with the same:

    Initializing: Operator is initializing

  reason/message pair I'm using for Progressing.

[1]: https://github.com/openshift/library-go/blob/94c59dec54be25c8527e51e8c0a885712aeb01b5/pkg/operator/status/condition.go#L67