Add healthchecks for Machine, MachineSet and MHC controllers
OCPCLOUD-785 - health checks for all machine API controllers

This introduce support for readinessProbe and livenessProbe [1] for the owned machine controllers deployment and its machineSet, MHC and machine controller containers.
This will let the kubelet to better acknowledge about these containers lifecycle and therefore letting us to be more robust to signal the operator degradability on the clusterOperator status.

This PR needs [2], [3] and [4] to work and pass CI so the probes included in the container spec here can get a 200 from the machine controllers. Also additionally [5], [6] and [7] must the same to not break or the probes included in the container spec here will fail will and result in the containers getting restarted.

This also reverts accidental rebase and put back the syncPeriod which was dropped by [8].

[1] https://kubernetes.io/docs/concepts/workloads/pods/pod-lifecycle/#container-probes
[2] openshift/cluster-api-provider-aws#329
[3] openshift/cluster-api-provider-azure#139
[4] openshift/cluster-api-provider-gcp#96
[5] https://github.com/openshift/cluster-api-provider-openstack
[6] https://github.com/openshift/cluster-api-provider-ovirt
[7] https://github.com/metal3-io/cluster-api-provider-metal3
[8] https://github.com/openshift/machine-api-operator/pull/590/files#diff-7417e4bc31a1bacc1a431704bee56978L41