Add support for optional provisioning networks in a Baremetal platform
In addition to supporting fully managed provisioning networks with the
option of having external DHCP servers, baremetal installations should
also be able to support cases where a dedicated provisioning network
does not exist.
Details regarding the updates to the provisioning CR to accomodate the
above feature can be found in [1].
Also, removing support for the ConfigMap as way to pass provisioning
configuration to the MAO for baremetal installs.

[1] - https://github.com/openshift/enhancements/blob/master/enhancements/baremetal/baremetal-provisioning-optional.md