add trusted ca bundle to machine-controllers
The machine-controllers communicate with AWS APIs that are outside the openshift cluster, therefore
it is possible that if the user has setup a Man-In-Middle-Terminating SSL Proxy for external traffic for reasons like auditing,
the machine-controllers will fail to communicate with AWS APIs as they do not trust the mimt-ssl-proxy's certificate.

The goal is to support just the case of mimt-proxy because this usually is applied using traffic shaping that directs all external traffic to it,
and therefore it makes sense to support it and still keep the previous stance that openshift expects to reach out to cloud APIs directly.

The implementation is very similar to the other cluster operators like cluster-ingress-operator in https://github.com/openshift/cluster-ingress-operator/pull/334/files

It creates a configmap in operator namespace with the annotation `config.openshift.io/inject-trusted-cabundle: "true"` which will include the complete trust for the container
and will be updated when user changes their configuration to add or change the user-defined trust for mimt-proxy.