operator: rollout the machine controller deployment on trust bundle change
So the trust bundle is loaded into the machine-controller's deployment as a config map volume, so the contents of the file
will automatically be syned when the config map contents change due to external factors.

But all the http clients that have already been created by the machine-controller like the cloud SDK clients, these will not see the
new trust unless re-created. So therefore to reload the trust for these clients we have 2 options,

- update all the cluster-api-provider implementations to include an ability to reload/restart when a file changes on disk.
  This requires a lot of work in each of the implementations and also creates a lot of duplicated code when we cannot use the same
  code for the behavior because we use upstream-downstream patches.

- track external dependencies and update the deployment object to rollout the changes as part of management in the machine-api-operator itself.
  This keeps the code to track this and others if we do end up extending it in one place in the operator. This method is extensively used
  by various other library-go based operators like openshift-apiserver-oeprator here https://github.com/openshift/cluster-openshift-apiserver-operator/blob/master/pkg/operator/workload/workload_openshiftapiserver_v311_00_sync.go#L340-L351
  and many others for similar behavior.