Add event handler for the daemonsetInformer which managed the termination handler
This https://github.com/openshift/machine-api-operator/pull/535 introduced support to manage a damonSet which runs termination handler for spot intances.
As an event handler is not passed to the damonSet informer changes to the resource won't trigger a reconcile.
This PR fix that by passing the event handler to the daemonSet namespaced informer.
This will be e2e tested by https://github.com/openshift/cluster-api-actuator-pkg/pull/177