baremetal: Deploy ironic with HTTP basic_auth
Pass the credentials as environment variables as expected by:
metal3-io/baremetal-operator#601
metal3-io/ironic-image#168
metal3-io/ironic-inspector-image#59

Note that this requires using the same credentials for both ironic-api
and the json-rpc interface to ironic-conductor.

Hard-code the username to "metal3", and use the same credentials for
both ironic and ironic-inspector containers.