Set default clusterID labels only on machine
This tries to fix the following scenario:
We set ms.Spec.Selector.MatchLabels[MachineClusterIDLabel] if it's not present. It's not present and we set it to the correct value. If there happens to be a bad label in `ms.Spec.Template.Labels` this would result in a miss match.

Follow for
https://github.com/openshift/machine-api-operator/pull/608,
https://github.com/openshift/machine-api-operator/pull/644 and
https://github.com/openshift/machine-api-operator/pull/653.