baremetal: Generate password hash for basic auth login
Services that are running with HTTP basic auth enabled need only the
password hash, not the raw password, to be able to authenticate
connections.

Pass the username and hash in the htpasswd format used by Ironic
services to authenticate connections instead of the raw password.

This relies on changes made in metal3-io/ironic-image#182 and
metal3-io/ironic-inspector-image#61.