baremetal: Create cryptographically secure passwords
Using a pseudorandom number generator seeded with the current time is a
not a secure way of generating a password. In this case the password is
generated immediately before storing it in a Secret - whose creation
time is known to at least a 1s resolution and is relatively accessible -
giving an effective entropy of at most around 30 bits, and quite
possibly (a lot) less.

Use a cryptographically secure pseudorandom number generator instead to
generate passwords.