Default MachineSet replicas to 1
The docs say that this should be defaulted to 1. The controller expects 
it to be non nil and errors if the value is nil. By defaulting to 1 we 
may break a small number of users who may have been using a nil value to 
prevent the machineset controller from working, but this is not an 
intended behaviour anyway.