Don't collect machine metrics if machine-controller not running
Currently, mapi_machine_created_timestamp_seconds is always
collected from the MAO.  On some UPI clusters, machine objects
may be created, even though the machine-controller is not
running.  This results in alerting for machines that are
missing the phase 'Running'.

This commit changes the MAO to only collect metrics for
machines that have non-empty, non-nil phase.