Ensure that negative values for maxUnhealthy do not produce events
Previously, we were producing a remediation restricted event every 
reconcile if a maxUnhealthy value is negative.
We can only validate the maxUnhealthy integer value by creating a 
webhook which is a substantial amount of work, this is a quicker fix 
that makes the behaviour of a negative value identical to that of a 0 
value.