MAO to stop managing metal3 deployments if CBO is available
In an upgrade scenario, check if CBO is available to take ownership
of the metal3 deployment. In that case, back off and let CBO manage it.

In a downgrade/rollback scenario where the metal3 deployment is
annotated as being owned by cluster-baremetal-operator, but the
baremetal operator has been removed, take back control of the metal3
deployment.