Don't use external remediation on failed Machines
Currently baremetal Machines never go into the Failed phase, but in the
future when they do it will be because of an error that is
unrecoverable: invalid config, unable to find a Host to provision,
underlying Host was deleted, &c.

In these cases, rebooting will not help, so always remediate them by
deleting the Machine.