Ensure an event is emitted when remediation is skipped
To improve the user experience of the MHC, we should always emit an 
event when we are skipping remediation.

We used to have this for master Machines, though this was removed when 
we migrated to the controller owner check. This restores the behaviour, 
but with a new event that is more agnostic of the type of Machine.