Ensure Progressing condition is stable when no updates to apply
The progressing condition is intedned to imply that a new version of the 
operator's resources are being rolled out. Currently we set it to resync 
on every reconcile which leads to constant updates and the value of the 
condiiton being unstable.

This change ensures that the status is stable when a version upgrade is 
not progressing and that we don't update the status every time we sync.