vSphere: Fix double create
Currently, some unknown condition can cause the
actuator to double create the clone task.  The
suspected reason is an eventual consistency problem
in the vSphere API which results in the task
object returned by the API as being nil immediately
after creation.

This commit ensures we only ever create a single
clone task.