vSphere: create requeue delay
The vSphere provider is reliant upon creating
and storing a task ID to track the cloning of a
template.  Without the task ID, we have no
record of requesting a template to create a new
instance.

Currently, it appears that we may be racing
with the controller's client cache and getting
a stale object when reconciling after initially
creating the task.  Even though we successfully
patch the status object with the new task ID,
the next reconcile may be utilizing out-dated
information and result in a double-create.