vSphere: preserve taskID in local cache
After patching the machine object following the initial
create, the subsequent reconcile may have a stale machine
object.  Additionally, any transient k8s API erros might
result in the patch operation failing and there is no
way to recover that operation. Either of these
scenarios will result in a double creation event
and a leaked instance.

This commit ensures we persist the taskID from the
vSphere API in case of this situations and preserve
it to ensure we do not double create.

This commit also removes the RequeueAfterError for
successful creation events as it is ineffectual.