add modular document templates
This change add markdown documents to be used as templates when creating
new docs. It also includes a readme file with a brief description of
usage and links to external resources. It's worth noting that although
these templates are in markdown format, they contain some liquid[0] and
jekyll[1] references which will be used with a static content generator.

[0] https://shopify.github.io/liquid/
[1] https://jekyllrb.com/