vSphere: More consistently wrap errors
Inconsistent use of wrapping means that in some cases the context of the
original error is lost, as it is replaced by an entirely new error
rather than just wrapped.

Crucially, this meant that sometimes an error created with
InvalidMachineConfiguration() was replaced instead of wrapped, meaning
that the Machine controller would not see it and would not change the
phase to Failed. This was inconsistent with the other machine API
providers.