Ensure zero value of remediationsAllowed is printed
So long as the value has been non-zero before, this will ensure that the 
value is printed. If it has never been non-zero, it's very difficult for 
us to set the field without making it a pointer, which we would 
preferably not do as it complicates a lot of logic throughout the 
codebase.