add alert for machine with long deletion phase
This change adds the MachineNotYetDeleted alert for a machine
with phase "Deleting" for longer than 360 minutes, at "warning" severity.
Also updates the alerts documentation with new information.

The MachineWithNoRunningPhase alert has also been modified to exclude
machines with the "Deleting" phase to ensure that we do not double fire
an alert for this condition.

ref: https://issues.redhat.com/browse/OCPCLOUD-921