install/0000_90_machine-api-operator_04_alertrules: Drop MachineAPIOperatorDown
The cluster-version operator is responsible for complaining if the
machine-API operator's deployment is sad, so no need for the operator
to handle this directly (we end up doubling up if there's an issue).
This drops the alert which was added in 9b666d6f20 (Add a new alert
rule for reporting the MAO failure, 2019-08-23, #388).