add avg_over_time to the MachineNotYetDeleted alert
This change is adding the averaging function to help smooth out drops
that can occur when the MAO is redeployed or moved within a cluster. It
is possible for the deletion phase timestamp metric to occassionaly drop
out, when this happens it can upset the entire 6 hour window for the
metric. With this change we introduce a 15 min sliding average for the
metric, which should alleviate issues where the MAO is gone for up to <
15 minutes.

Also updates the alerts document.