[OCPCLOUD-1260] modify prometheus alerts to ignore pod
By default prometheus metrics in kubernetes contain labels for the
pod and instance that created a metric series. If these labels
change during the metric series then it causes a new metric series
to be created. When this happens, a long running alert will get
reset even if the target of the alert (eg a machine) is the same.

This has deleterious effects when considering things like machines
in "Deleting" phase, or missing a Node reference.

This change makes it so that the alert will ignore changes in the pod
label during the duration of the alert. It is being added to account for
cases where the machine-api-operator pod might be rescheduled and thus
its pod name will change. This should allow the alert to persist.