Don't force a requeue after updating Status.Phase
The forced requeue is redundant as we just updated Status.Phase so the
informer will call us again anyway when it sees the change.

Worse, forcing the requeue causes us to be called before the informer
sees the Status update, meaning we're called with a stale Machine
object. This will cause a storm of Status updates until the informer
catches up.