add utility functions for detecting powershell tags
This change adds a set of helper functions that will be used with
Windows machine deployments. When booting a Windows machine, it is
common for the infrastructure provider to have a method for a Powershell
script to run during boot. This is usually supplied in the user data
metadata associated with a new instance. On Azure and AWS, the script is
required to be wrapped in `<powershell></powershell>` tags, but on GCP
this is not required. The helper functions in this commit will be used
by the providers to help in detecting, adding, and removing these tags.