add more support for old delete annotation
this change adds detection to the newest and random delete priority
functions to ensure that we have coverage for the older annotation until
it is removed. also updates the tests to cover a situation where the old
annotation was not detected properly and adds more tests for the new
interactions.