update windows helper functions
this change updates the tag identifiers that are used by the WMCO when
creating the secret for the windows boot script. we need to have a
slightly tighter coupling with the WMCO until such time as they drop the
tags from their implementation. once the WMCO removes all the tags, we
should evualte the usefulness of the helper functions and remove or
modify them as necessary. for context, only some of the cloud provider
will need these tags, currently AWS and Azure.