OCPBUGS-2151: Don't degrade when workers not expected
Since 28853af469d6e3e6c40454de85baa7084d3165c0 we mark the MAO degraded
when two workers have not come up (and we ignore this constraint if zero
workers are expected), but this will never succeed if exactly one worker
has been requested. This was previously an invalid configuration, but
since MGMT-8987 it is permitted to create a compact cluster with exactly
one worker, since the control plane nodes are schedulable.

To handle this case, only remain degraded when we have fewer than two
workers *and* all of the expected workers have not yet come up. Leave it
to installers to validate which configurations are and are not supported
on Day 1.

This does mean that the operator can report that it is degraded even
when there are sufficient machines to run e.g. the Ingress operator due
to schedulable masters (which could be changed only by checking the
scheduler config). However, given that in IPI workers and masters are
provisioned by entirely different methods, it makes sense to report
degraded so long as any workers have been requested but none created.