Fix getNetworkDevice method of the vsphere machine reconciler
In case of multiple network devices in machine's
provider spec, it was possible to wrongly use device Backing info
from the previously handled network device instead of returning an error
in a situation when Network specified for the device does not exist.
This might lead to the creation of machines which does not comply with
the spec.

Now machine is supposed to go into the Failed phase
in case if the desired network is not found during machine cloning.