Forcefully delete unevicted pods within vSphere machine deletion procedure
In case the drain procedure was performed on a node where kubelet was not
operational for some reason, pods on a such node will be stuck in the "Terminating"
state and will be never deleted.
If such pods have volumes attached to them, these volumes will never be detached.

For the case when VM was already drained and powered off,
but still has attached volumes
we need to attempt forcefully delete pods in "Terminating" state.
This will implicitly trigger attach-detach controller to
detach disks from the VM.