Adds the MergeCommaSeparatedKeyValues method in the util package
In the efforts for MIXEDARCH-129, we will need to change several cloud providers' actuators to get the kubernetes.io/arch label considered for scaling from zero. The upstream changes introduced by kubernetes/autoscaler/pull/5382 allow us to use an annotation to instruct the autoscaler about the node labels. The actuators will consume the method introduced by this commit to merge the comma-separated list of key-value pairs stored in that annotation with any others the actuators want to add.