Add feature gate for disabling the MHC controller
There is a plan to let the Node Healthcheck Operator, which is an
optional day 2 operator today, also implement the MHC controller. For
being able to test the POC in NHC, we need to be able to disable MAO's
MHC controller. So for now this feature gate is just for dev usage.
The actual transition of the MHC controller from MAO to NHC will be
discussed and planned in the future.

Signed-off-by: Marc Sluiter <msluiter@redhat.com>