Per-architecture defaults on Azure
Introduce per-architecture defaults on Azure. Beyond the instance type, for arm64, we add the -gen2 postfix to the gallery name. In fact, since the arm64 enablement, we create an image gallery and images for the Gen2 HyperV. It is done for both x86_64 and aarch64. However, for x86_64 clusters installed before that change, we need yet to maintain the non -gen2 image gallery name as that might be the only one available.