handle nodes without UID
On a large deployment on OpenStack, with high latency, we observe the
mhc controller remediating nodes immediately after their creation.  The
remediation happens without any log for its cause, and the only path in
the code where that is the case is when the node.UID is empty.  We do
see a node Name in the target string() that is logged, so we know that
a node.Name is present at that time.

Change the code so that a node is considered to be existing when it has
a name and no UID, even if the circumstances by which this happens are
blurry.

Also add a log so that all code paths leading to remediation are clearly
seen.

Signed-off-by: François Rigault <frigo@amadeus.com>