Machine feature gating
This change adds feature gating to the Machine Controller. The interface
is the same as used by the MachineSet controller, where features are
passed as flags via --feature-gates=<name>=<bool>

This is a breaking change, as we change AddWithActuator to expect a
feature gate. This means each MAPI provider will need updating to do
this.