azure: remove image defaulting
Removes the default image value populated by the webhook. Uses existing
validation to ensure the image id is populated.

The installer switched to defaulting to marketplace images, and no
longer creates a cluster-specific image gallery (or managed image).
So the default value populated by the webhook is no longer correct.
The behavior now is similar to AWS, which requires an image be
specified.